Added initial CSS and changed HTML Structure
In order to get the basic layout in the goal image, I began adding some
CSS. I added background colors and some font colors.  In the process of
doing this I began changing some of the HTML structure to better fit
the goal image of the project. This included placing the hero in the
header, and setting up a new div for the quote section.