Added images, completed Header/hero section
Added and sized some images for the various sections, using an oceanic
theme because I think it fits well with the color palette of the page
and I enjoy nautical themes generally. Completed Header/Hero sections.
Required a lot of tinkering in CSS. Used margin-right: auto for
alignment purposes in both header and in content. Nested Flexboxes for
getting image and content placed correctly. I think there may have been
a simpler way to pull this off but I like that with flex it has a little
bit of responsiveness as the window gets resized.