Finished Call to Action Section
Call to action section complete. used nested flexboxes to get content
to align and behave similarly to goal image. One div (calltoact) was
used to set up the area for the content and was centered with flexbox.
Then set up a container div (container) to separate the text elements
from the button element and have them aligned correctly. Finally set up
a div for just the text elements (acttext) and set it up as a flexbox
column to match the goal image. Used appropriate CSS from the info.png
to match styling.